Move ObservableSet into a separate file
